+++ tests/testthat/test-validate-segmented.R
@@ -26,13 +26,14 @@ test_that("Gaussian 2-changepoint model agrees with segmented()", {
 
   # Extract mcp estimates
   fix_mcp = fixef(fit_mcp)
-  cp1_mcp = fix_mcp$mean[fix_mcp$name == "cp_1"]
-  cp2_mcp = fix_mcp$mean[fix_mcp$name == "cp_2"]
+  summary_mcp = summary(fit_mcp)
+  cp1_mcp = summary_mcp$mean[summary_mcp$name == "cp_1"]
+  cp2_mcp = summary_mcp$mean[summary_mcp$name == "cp_2"]
   intercept_mcp = fix_mcp$mean[fix_mcp$name == "Intercept_1"]
   x1_mcp = fix_mcp$mean[fix_mcp$name == "x_1"]
   x2_mcp = fix_mcp$mean[fix_mcp$name == "x_2"]
   x3_mcp = fix_mcp$mean[fix_mcp$name == "x_3"]
-  sigma_mcp = fix_mcp$mean[fix_mcp$name == "sigma_1"]
+  sigma_mcp = summary_mcp$mean[summary_mcp$name == "sigma_1"]
 
   # Extract segmented estimates
   psi_seg = fit_seg$psi[, "Est."]
@@ -76,8 +77,9 @@ test_that("Poisson 2-changepoint model agrees with segmented()", {
 
   # Extract mcp estimates
   fix_mcp = fixef(fit_mcp)
-  cp1_mcp = fix_mcp$mean[fix_mcp$name == "cp_1"]
-  cp2_mcp = fix_mcp$mean[fix_mcp$name == "cp_2"]
+  summary_mcp = summary(fit_mcp)
+  cp1_mcp = summary_mcp$mean[summary_mcp$name == "cp_1"]
+  cp2_mcp = summary_mcp$mean[summary_mcp$name == "cp_2"]
   intercept_mcp = fix_mcp$mean[fix_mcp$name == "Intercept_1"]
   x1_mcp = fix_mcp$mean[fix_mcp$name == "x_1"]
   x2_mcp = fix_mcp$mean[fix_mcp$name == "x_2"]
@@ -124,8 +126,9 @@ test_that("Binomial 2-changepoint model agrees with segmented()", {
 
   # Extract mcp estimates
   fix_mcp = fixef(fit_mcp)
-  cp1_mcp = fix_mcp$mean[fix_mcp$name == "cp_1"]
-  cp2_mcp = fix_mcp$mean[fix_mcp$name == "cp_2"]
+  summary_mcp = summary(fit_mcp)
+  cp1_mcp = summary_mcp$mean[summary_mcp$name == "cp_1"]
+  cp2_mcp = summary_mcp$mean[summary_mcp$name == "cp_2"]
   intercept_mcp = fix_mcp$mean[fix_mcp$name == "Intercept_1"]
   x1_mcp = fix_mcp$mean[fix_mcp$name == "x_1"]
   x2_mcp = fix_mcp$mean[fix_mcp$name == "x_2"]